+++ plugins/wazuh-core/common/constants.ts
@@ -586,28 +586,31 @@ export const PLUGIN_SETTINGS: Record<string, TPluginSetting> = {
     type: EpluginSettingType.switch,
     validate: SettingsValidator.isBoolean,
   },
-  'enrollment.dns': {
-    title: 'Enrollment DNS',
-    description:
-      'Specifies the Wazuh registration server, used for the agent enrollment.',
+  'enrollment.url': {
+    title: 'Enrollment server URL',
+    description: 'Specifies the enrollment agent server URL.',
     source: EConfigurationProviders.PLUGIN_UI_SETTINGS,
     category: SettingCategory.GENERAL,
     type: EpluginSettingType.text,
     defaultValue: '',
-    validate: SettingsValidator.compose(
-      SettingsValidator.isString,
-      SettingsValidator.serverAddressHostnameFQDNIPv4IPv6,
-    ),
+    // TODO: this should be enabled when the configuration service of core plugin provides a mechanism to retrieve this definition to be used in the enrollment agent wizard. See https://github.com/wazuh/wazuh-dashboard/issues/514#issuecomment-2656602679
+    // validate: SettingsValidator.compose(
+    //   SettingsValidator.isString,
+    //   SettingsValidator.serverAddressHostnameFQDNIPv4IPv6,
+    // ),
   },
-  'enrollment.password': {
-    title: 'Enrollment password',
-    description:
-      'Specifies the password used to authenticate during the agent enrollment.',
+  'enrollment.commsUrl': {
+    title: 'Enrollment communications URL',
+    description: 'Specifies the enrollment communications API URL.',
     source: EConfigurationProviders.PLUGIN_UI_SETTINGS,
     category: SettingCategory.GENERAL,
     type: EpluginSettingType.text,
     defaultValue: '',
-    validate: SettingsValidator.compose(SettingsValidator.isString),
+    // TODO: this should be enabled when the configuration service of core plugin provides a mechanism to retrieve this definition to be used in the enrollment agent wizard. See https://github.com/wazuh/wazuh-dashboard/issues/514#issuecomment-2656602679
+    // validate: SettingsValidator.compose(
+    //   SettingsValidator.isString,
+    //   SettingsValidator.serverAddressHostnameFQDNIPv4IPv6,
+    // ),
   },
   hideManagerAlerts: {
     title: 'Hide manager alerts',
@@ -1020,7 +1023,6 @@ export const CRON_PREFIX = 'cron.prefix';
 export const CUSTOMIZATION_ENABLED = 'customization.enabled';
 
 export const ENROLLMENT_DNS = 'enrollment.dns';
-export const ENROLLMENT_PASSWORD = 'enrollment.password';
 
 export const IP_IGNORE = 'ip.ignore';
 export const IP_SELECTOR = 'ip.selector';