+++ plugins/wazuh-core/public/types.ts
@@ -1,4 +1,8 @@
 import React from 'react';
+import {
+  DataPublicPluginSetup,
+  DataPublicPluginStart,
+} from 'src/plugins/data/public';
 import { API_USER_STATUS_RUN_AS } from '../common/api-user-status-run-as';
 import { Configuration } from '../common/services/configuration';
 import { State, StateSetupReturn } from './services/state';
@@ -13,6 +17,14 @@ import {
   DashboardSecurityServiceSetupReturn,
 } from './services/dashboard-security';
 import { ApplicationService } from './services/application/application';
+
+export interface AppPluginSetupDependencies {
+  data: DataPublicPluginSetup;
+}
+
+export interface AppPluginStartDependencies {
+  data: DataPublicPluginStart;
+}
 
 export interface WazuhCorePluginSetup {
   _internal: any;
@@ -71,5 +83,3 @@ export interface WazuhCorePluginStart {
     ) => React.ComponentType<ServerDataProps<T>>;
   } & ServerSecuritySetupReturn['ui'];
 }
-
-export type AppPluginStartDependencies = object;