+++ plugins/wazuh-fleet/public/application/types.ts
@@ -1,0 +1,11 @@
+export interface AppSetup {
+  registerApp: (app: any) => void;
+  enrollmentAgentManagement: {
+    serverURLSettingName: string;
+    getServerURL: () => Promise<string>;
+    setServerURL: (url: string) => Promise<string>;
+    commsURLSettingName: string;
+    getCommunicationsURL: () => Promise<string>;
+    setCommunicationsURL: (url: string) => Promise<string>;
+  };
+}